# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Agrafa backend API for agent ingestion, inventory creation, and read-side observability queries.",
        "title": "Agrafa API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/services": {
            "get": {
                "description": "Returns newest services first, with execution_mode derived from the attached node and optional project_id, node_id, status, and limit filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "List services",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Node ID",
                        "name": "node_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "healthy",
                            "degraded",
                            "unhealthy"
                        ],
                        "type": "string",
                        "description": "Service status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Maximum number of services",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ServicesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a service under a project and resolves the execution node from execution_mode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Create service",
                "parameters": [
                    {
                        "description": "Service payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ServiceCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.ServiceDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "description": "Returns service details for a service the current user can read.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Get service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ServiceDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a service and relies on existing schema cleanup for dependent operational records.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Delete service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the editable service definition fields only. Execution mode and attached node are unchanged.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Update service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ServiceUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ServiceDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "types.HealthCheckSummaryDocument": {
            "type": "object",
            "properties": {
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "observed_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "response_time_ms": {
                    "type": "integer",
                    "example": 42
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "types.ServiceCreateRequest": {
            "type": "object",
            "properties": {
                "check_target": {
                    "type": "string",
                    "example": "https://api.planzi.app/health"
                },
                "check_type": {
                    "type": "string",
                    "example": "http"
                },
                "execution_mode": {
                    "type": "string",
                    "example": "managed"
                },
                "name": {
                    "type": "string",
                    "example": "planzi-api"
                },
                "node_id": {
                    "type": "integer",
                    "example": 1
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "types.ServiceDetailResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "$ref": "#/definitions/types.ServiceDocument"
                }
            }
        },
        "types.ServiceDocument": {
            "type": "object",
            "properties": {
                "active_alert_count": {
                    "type": "integer",
                    "example": 1
                },
                "check_target": {
                    "type": "string",
                    "example": "https://api.planzi.app/health"
                },
                "check_type": {
                    "type": "string",
                    "example": "http"
                },
                "consecutive_failures": {
                    "type": "integer",
                    "example": 0
                },
                "execution_mode": {
                    "type": "string",
                    "example": "agent"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_checked_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "latest_health_check": {
                    "$ref": "#/definitions/types.HealthCheckSummaryDocument"
                },
                "name": {
                    "type": "string",
                    "example": "planzi-api"
                },
                "node_id": {
                    "type": "integer",
                    "example": 1
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                }
            }
        },
...
        "types.ServiceUpdateRequest": {
            "type": "object",
            "properties": {
                "check_target": {
                    "type": "string",
                    "example": "https://api.planzi.app/status"
                },
                "check_type": {
                    "type": "string",
                    "example": "http"
                },
                "name": {
                    "type": "string",
                    "example": "planzi-api-v2"
                }
            }
        },
        "types.ServicesResponse": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ServiceDocument"
                    }
                }
            }
        },
...
    }
}